CRUD UNIVERSITE

+++ back/src/main/java/com/example/springproject/Repository/UniversiteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversiteRepository extends CrudRepository<Universite,Long> {
+public interface UniversiteRepository extends CrudRepository<Universite,Integer> {
 }
